feat: ship a consumer-runnable v6 attestation verifier (AISDLC-566)
AISDLC-554 made the DSSE signer reachable in adopter/consumer repos via a
candidate-walk runtime resolver, but the verifier side was untouched:
scripts/verify-attestation.mjs (the pass/fail verifier this repo's CI uses)
was never shipped in the plugin and had a static monorepo-relative import
that only resolves inside this checkout.

Extracts the full v6/v5/v4/v3 verification core (Merkle primitives,
head-binding relaxations, content-hash matching, runVerifier) out of
scripts/verify-attestation.mjs into a shared module,
ai-sdlc-plugin/scripts/verify-attestation-core.mjs, imported by BOTH:

- scripts/verify-attestation.mjs — now a thin driver, unchanged behaviour,
  keeps .github/workflows/verify-attestation.yml working.
- ai-sdlc-plugin/scripts/verify-attestation.mjs — new consumer-runnable
  driver that resolves @ai-sdlc/orchestrator via the same candidate-walk
  strategy sign-attestation.mjs uses (published package export /
  node_modules walk-up / plugin dirs, version-gated).

Adds docs/operations/adopter-attestation-verify-ci.md (copy-pasteable
GitHub Actions recipe) and a hermetic test suite proving real sign-then-
verify round trips (v5 and v6) against a node_modules-style install layout
with no orchestrator/ source tree present.
