+++ package.json
@@ -12,7 +12,7 @@
   "scripts": {
     "typecheck": "pnpm -r --parallel exec tsc --noEmit",
     "build": "pnpm -r build",
-    "test": "pnpm -r test && pnpm docs:test && pnpm docs:check && pnpm test:publishable && pnpm rfc:test && pnpm rfc:check && pnpm test:drift-gate && pnpm test:dark-code-gate && pnpm dark-code:check && pnpm test:task-move-gate && pnpm test:attestation-sign-gate && pnpm test:plugin-attestation-sign-gate && pnpm test:sign-attestation-gate && pnpm test:install-runtime-deps-gate && pnpm test:verify-attestation-gate && pnpm test:orchestrator-state-gate && pnpm test:gate-workflow && pnpm test:review-workflow && pnpm test:fork-pr-safety && pnpm test:verify-attestation-workflow && pnpm test:docs-only-classifier && pnpm test:audit-gate && pnpm test:drop-stale-envelope && pnpm test:auto-rebase-stale-prs && pnpm test:dor-gate && pnpm test:pre-push-fixups-gate && pnpm test:adopter-facing-strings && pnpm test:patch-coverage-gate && pnpm test:coverage-gate && pnpm test:release-please-config && pnpm test:require-issue-link && pnpm test:supply-chain-hardening && pnpm test:substrate-contract-gate && pnpm test:substrate-contract-conformance && pnpm test:execute-parallel-spawn-gate && pnpm test:backlog-drift-push-gate && pnpm test:untrusted-pr-gate-workflow",
+    "test": "pnpm -r test && pnpm docs:test && pnpm docs:check && pnpm test:publishable && pnpm rfc:test && pnpm rfc:check && pnpm test:drift-gate && pnpm test:dark-code-gate && pnpm dark-code:check && pnpm test:task-move-gate && pnpm test:attestation-sign-gate && pnpm test:plugin-attestation-sign-gate && pnpm test:sign-attestation-gate && pnpm test:install-runtime-deps-gate && pnpm test:verify-attestation-gate && pnpm test:verify-attestation-plugin-gate && pnpm test:orchestrator-state-gate && pnpm test:gate-workflow && pnpm test:review-workflow && pnpm test:fork-pr-safety && pnpm test:verify-attestation-workflow && pnpm test:docs-only-classifier && pnpm test:audit-gate && pnpm test:drop-stale-envelope && pnpm test:auto-rebase-stale-prs && pnpm test:dor-gate && pnpm test:pre-push-fixups-gate && pnpm test:adopter-facing-strings && pnpm test:patch-coverage-gate && pnpm test:coverage-gate && pnpm test:release-please-config && pnpm test:require-issue-link && pnpm test:supply-chain-hardening && pnpm test:substrate-contract-gate && pnpm test:substrate-contract-conformance && pnpm test:execute-parallel-spawn-gate && pnpm test:backlog-drift-push-gate && pnpm test:untrusted-pr-gate-workflow",
     "test:coverage": "pnpm -r test:coverage",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
@@ -37,6 +37,7 @@
     "test:install-runtime-deps-gate": "node --test ai-sdlc-plugin/scripts/install-runtime-deps.test.mjs",
     "test:execute-parallel-spawn-gate": "node --test ai-sdlc-plugin/scripts/execute-parallel-spawn.test.mjs",
     "test:verify-attestation-gate": "node --test scripts/verify-attestation.test.mjs",
+    "test:verify-attestation-plugin-gate": "node --test ai-sdlc-plugin/scripts/verify-attestation.test.mjs",
     "test:orchestrator-state-gate": "node --test scripts/check-orchestrator-state.test.mjs",
     "test:gate-workflow": "node --test .github/workflows/__tests__/ai-sdlc-gate.test.mjs",
     "test:review-workflow": "node --test .github/workflows/__tests__/ai-sdlc-review.test.mjs",
